+++ src/app/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React from 'react';
-import { AppContent } from '@/components/AppContent';
+import { AppContent } from '@/components/App';
 
 export default function Home() {
   return (
