--- src/components/LabelFilter.module.css
@@ -1,97 +1,0 @@
-.labelFilter {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  margin-bottom: 20px;
-  width: 100%;
-}
-
-.selectedLabels {
-  display: flex;
-  gap: 5px;
-  flex-wrap: wrap;
-}
-
-.label {
-  background-color: #3e3e3e;
-  color: #e0e0e0;
-  padding: 2px 5px;
-  border-radius: 3px;
-  font-size: 0.8em;
-  display: flex;
-  align-items: center;
-}
-
-.removeButton {
-  background: none;
-  border: none;
-  color: #cf6679;
-  margin-left: 5px;
-  padding: 0 2px 0 0;
-  cursor: pointer;
-}
-
-.removeButton:hover {
-  color: #ff8a80;
-}
-
-.comboboxInput {
-  background-color: #3e3e3e;
-  color: #e0e0e0;
-  border: 1px solid #555;
-  padding: 5px;
-  border-radius: 3px;
-}
-
-.comboboxInput::placeholder {
-  color: #aaa;
-}
-
-.comboboxList {
-  background-color: #3e3e3e;
-  color: #e0e0e0;
-  border: 1px solid #555;
-  border-radius: 3px;
-}
-
-.comboboxOption {
-  padding: 5px;
-}
-
-.comboboxOption:hover {
-  background-color: #555;
-}
-
-.comboboxContainer {
-  display: flex;
-  width: 100%;
-}
-
-.comboboxContainer div, .comboboxContainer input {
-  width: 100%;
-}
-
-.addButton {
-  margin-left: 8px;
-  padding: 4px 8px;
-  border: none;
-  cursor: pointer;
-}
-
-.addButton:hover {
-  background-color: #0056b3;
-}
-
-.token {
-  max-width: 200px;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: normal;
-  word-wrap: break-word;
-}
-
-.tokenContainer {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 5px;
-}
