+++ src/components/Token.module.css
@@ -1,16 +1,37 @@
 .token {
-  display: inline-flex;
-  align-items: center;
-  padding: 0.5em;
-  margin: 0.5em;
-  background-color: #e0e0e0;
-  border-radius: 0.25em;
+  max-width: 200px;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: normal;
+  word-wrap: break-word;
 }
 
+.tokenContainer {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 5px;
+}
+
+
 .removeButton {
-  margin-left: 0.5em;
   background: none;
   border: none;
+  color: #cf6679;
+  margin-left: 5px;
+  padding: 0 2px 0 0;
   cursor: pointer;
-  color: #ff0000;
+}
+
+.removeButton:hover {
+  color: #ff8a80;
+}
+
+.label {
+  background-color: #3e3e3e;
+  color: #e0e0e0;
+  padding: 2px 5px;
+  border-radius: 3px;
+  font-size: 0.8em;
+  display: flex;
+  align-items: center;
 }
