+++ src/components/filters/ReasonFilters.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React, { useContext } from 'react';
-import styles from './AdditionalFilters.module.css';
+import styles from './ReasonFilters.module.css';
 import { AppContext, AppDispatchContext } from '@/store/AppContext';
 import { NotificationReason, ReasonFilter } from '@/types';
 import { isParticipating } from '@/shared/filterHelpers';
@@ -21,15 +21,10 @@ const SelectFilter: React.FC<{ filter: ReasonFilter | null, label: string, count
   );
 };
 
-const AdditionalFilters = () => {
-  const { stateFilter, notifications } = useContext(AppContext);
-  const dispatch = useContext(AppDispatchContext);
+const ReasonFilters = () => {
+  const { notifications } = useContext(AppContext);
 
-  const handleStateChange = (event: React.ChangeEvent<HTMLSelectElement>) => {
-    dispatch({ type: "SET_STATE_FILTER", payload: event.target.value }); // Call dispatch
-  };
-
-  const filterCounts: Record<ReasonFilter | NotificationReason, number> = notifications.reduce((acc, notification) => {
+  const filterCounts = notifications.reduce((acc, notification) => {
     const _isParticipating = isParticipating(notification.reason)
     const reason = notification.reason
 
@@ -39,7 +34,7 @@ const AdditionalFilters = () => {
       acc['participating'] = (acc['participating'] || 0) + 1;
     }
     return acc;
-  }, {} as Record<string, number>);
+  }, {} as Record<ReasonFilter | NotificationReason, number>);
 
 
   return (
@@ -50,17 +45,8 @@ const AdditionalFilters = () => {
       <SelectFilter filter="mention" label="Mentioned" count={filterCounts['mention'] || 0} />
       <SelectFilter filter="team_mention" label="Team Mentioned" count={filterCounts['team_mention'] || 0} />
       <SelectFilter filter="review_requested" label="Review Requested" count={filterCounts['review_requested'] || 0} />
-
-      <div>
-        <label htmlFor="stateFilter">State:</label>
-        <select id="stateFilter" value={stateFilter} onChange={handleStateChange}>
-          <option value="all">All</option>
-          <option value="open">Open</option>
-          <option value="closed">Closed</option>
-        </select>
-      </div>
     </div>
   );
 };
 
-export default AdditionalFilters;
+export default ReasonFilters
