+++ src/components/settings/SettingsPane.tsx
@@ -5,8 +5,8 @@ import { createPortal } from 'react-dom';
 import LabelFilter from './LabelFilter';
 import RepoPrioritization from './RepoPrioritization';
 import styles from './SettingsPane.module.css';
-import { GearIcon, CloseIcon } from './icons';
-import RateLimit from './RateLimit';
+import { GearIcon, CloseIcon } from '../icons';
+import RateLimit from '../RateLimit';
 
 const SettingsPane = () => {
     const [isVisible, setIsVisible] = useState(false);
